Add compatibility object for 0.13 and 1.0
